Fix svgPath issue for paths with spacing
Currently when a svgPath has a space (for example, `C://owo/John Doe`). The command is executed as `svgr C://owo/John Doe ...`, resulting in the file not being found since it believes the location is "C://owo/John". Hence, having double quotes to preserve the spaces in the path